% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobit.list.R
\docType{data}
\name{tobit.list}
\alias{tobit.list}
\title{Data: A \code{mcmc.list} from a Tobit model}
\description{
The data is the output from a Tobit model that was estimated with MCMC techniques.
	  The estimation was performed with five parallel chains of 50,000 iterations
	  each and the first 40,000 iterations were discarded.
}
\details{
The R code for the Gibbs sampler is available at \url{https://jeffgill.org/}.
}
\references{
Gill, Jeff. (2007). \emph{Bayesian Methods for the Social and Behavior
  Sciences.} Second Edition. New York: Chapman & Hall.
}
\keyword{data}
