\name{tobit.list}
\alias{tobit.list} 
\title{tobit.list}

\description{
	A Tobit Model was estimated with MCMC techniques. The estimation was performed with five parallel chains of 50,000 iterations each and the first 40,000 iterations were discarded.
}

\usage{data(tobit.list)}

\details{
	The R code for the Gibbs sampler is given at \url{http://jgill.wustl.edu/computing.html}.
}

\references{
	Gill, Jeff. (2007). \emph{Bayesian Methods for the Social and Behavior Sciences.} Second Edition. New York: Chapman & Hall.
}

\keyword{datasets}

